authentication implemented with email functionality

+++ controllers/admin.js
@@ -62,6 +62,9 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(prodId)
     .then(product => {
+      if(product.userId.toString() !== req.user._id) {
+          return res.redirect('/admin/products');
+      }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
@@ -76,7 +79,7 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find()
+  Product.find({userId: req.user._id})
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
@@ -92,10 +95,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndRemove(prodId)
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => console.log(err));
+    Product.findById(prodId)
+        .then(product => {
+            if(product.userId.toString() !== req.user._id) {
+                return res.redirect('/admin/products');
+            }
+            Product.findByIdAndRemove(prodId)
+                .then(() => {
+                    console.log('DESTROYED PRODUCT');
+                    res.redirect('/admin/products');
+                })
+                .catch(err => console.log(err));
+        })
+        .catch(err => console.log(err));
 };
